Separate XR components into xr namespace
